Add Create Article Functionality
Add isPublished Data for MyArticleCard 
Now article can be created within a dialog inside profile page.

The created articles will be set to isPublished:false to edit and publish users should be directed to /articles/:slug/edit

in the edit page they will be a publish/update button depending on the state.